+++ vimgolf/core.py
@@ -1,9 +1,7 @@
 import datetime
-import filecmp
 import json
 import os
 import re
-import subprocess
 import sys
 import urllib.parse
 import urllib.request
@@ -16,20 +14,10 @@ from vimgolf import (
     EXPANSION_PREFIX,
     GOLF_HOST,
     VIMGOLF_CHALLENGES_PATH,
-    GOLF_VIM,
-    PLAY_VIMRC_PATH,
-)
-from vimgolf.keys import (
-    IGNORED_KEYSTROKES,
-    get_keycode_repr,
-    parse_keycodes,
 )
 from vimgolf.utils import (
     write,
     http_request,
-    find_executable,
-    confirm,
-    input_loop,
 )
 
 
@@ -238,165 +226,3 @@ def upload_result(challenge_id, api_key, raw_keys):
     except Exception:
         logger.exception('upload failed')
     return status
-
-
-def play(challenge, workspace):
-    logger.info('play(...)')
-
-    vim_path = find_executable(GOLF_VIM)
-    if not vim_path:
-        write('Unable to find "{}"'.format(GOLF_VIM), color='red')
-        write('Please update your PATH to include the directory with "{}"'.format(GOLF_VIM), color='red')
-        return Status.FAILURE
-    vim_name = os.path.basename(os.path.realpath(vim_path))
-
-    if sys.platform == 'win32':
-        # Remove executable extension (.exe, .bat, .cmd, etc.) from 'vim_name'
-        base, ext = os.path.splitext(vim_name)
-        pathexts = os.environ.get('PATHEXT', '.EXE').split(os.pathsep)
-        for pathext in pathexts:
-            if ext.upper() == pathext.upper():
-                vim_name = base
-                break
-
-    # As of 2019/3/2, on Windows, nvim-qt doesn't support --nofork.
-    # Issue a warning as opposed to failing, since this may change.
-    if vim_name == 'nvim-qt' and sys.platform == 'win32':
-        write('vimgolf with nvim-qt on Windows may not function properly', color='red')
-        write('If there are issues, please try using a different version of vim', color='yellow')
-        if not confirm('Continue trying to play?'):
-            return Status.FAILURE
-
-    def vim(args, **run_kwargs):
-        # Configure args used by all vim invocations (for both playing and diffing)
-        # 'vim_path' is used instead of GOLF_VIM to handle 'vim.bat' on the PATH.
-        # subprocess.run would not launch vim.bat with GOLF_VIM == 'vim', but 'find_executable'
-        # will return the full path to vim.bat in that case.
-        vim_args = [vim_path]
-        # Add --nofork so gvim, mvim, and nvim-qt don't return immediately
-        # Add special-case handling since nvim doesn't accept that option.
-        if vim_name != 'nvim':
-            vim_args.append('--nofork')
-        # For nvim-qt, options after '--' are passed to nvim.
-        if vim_name == 'nvim-qt':
-            vim_args.append('--')
-        vim_args.extend(args)
-        subprocess.run(vim_args, **run_kwargs)
-        # On Windows, vimgolf freezes when reading input after nvim's exit.
-        # For an unknown reason, shell'ing out an effective no-op works-around the issue
-        if vim_name == 'nvim' and sys.platform == 'win32':
-            os.system('')
-
-    infile = os.path.join(workspace, 'in')
-    if challenge.in_extension:
-        infile += challenge.in_extension
-    outfile = os.path.join(workspace, 'out')
-    if challenge.out_extension:
-        outfile += challenge.out_extension
-    logfile = os.path.join(workspace, 'log')
-    with open(outfile, 'w') as f:
-        f.write(challenge.out_text)
-
-    write('Launching vimgolf session', color='yellow')
-    while True:
-        with open(infile, 'w') as f:
-            f.write(challenge.in_text)
-
-        vimrc = PLAY_VIMRC_PATH
-        play_args = [
-            '-Z',           # restricted mode, utilities not allowed
-            '-n',           # no swap file, memory only editing
-            '--noplugin',   # no plugins
-            '-i', 'NONE',   # don't load .viminfo (e.g., has saved macros, etc.)
-            '+0',           # start on line 0
-            '-u', vimrc,    # vimgolf .vimrc
-            '-U', 'NONE',   # don't load .gvimrc
-            '-W', logfile,  # keylog file (overwrites existing)
-            infile,
-        ]
-        try:
-            vim(play_args, check=True)
-        except Exception:
-            logger.exception('{} execution failed'.format(GOLF_VIM))
-            write('The execution of {} has failed'.format(GOLF_VIM), stream=sys.stderr, color='red')
-            return Status.FAILURE
-
-        correct = filecmp.cmp(infile, outfile)
-        with open(logfile, 'rb') as _f:
-            # raw keypress representation saved by vim's -w
-            raw_keys = _f.read()
-
-        # list of parsed keycode byte strings
-        keycodes = parse_keycodes(raw_keys)
-        keycodes = [keycode for keycode in keycodes if keycode not in IGNORED_KEYSTROKES]
-
-        # list of human-readable key strings
-        keycode_reprs = [get_keycode_repr(keycode) for keycode in keycodes]
-
-        score = len(keycodes)
-
-        write('Here are your keystrokes:', color='green')
-        for keycode_repr in keycode_reprs:
-            color = 'magenta' if len(keycode_repr) > 1 else None
-            write(keycode_repr, color=color, end=None)
-        write('')
-
-        if correct:
-            write('Success! Your output matches.', color='green')
-            write('Your score:', color='green')
-        else:
-            write('Uh oh, looks like your entry does not match the desired output.', color='red')
-            write('Your score for this failed attempt:', color='red')
-        write(score)
-
-        uploaded = False
-        upload_eligible = challenge.id and challenge.compliant and challenge.api_key
-
-        while True:
-            # Generate the menu items inside the loop since it can change across iterations
-            # (e.g., upload option can be removed)
-            menu = []
-            if not correct:
-                menu.append(('d', 'Show diff'))
-            if upload_eligible and correct:
-                menu.append(('w', 'Upload result'))
-            menu.append(('r', 'Retry the current challenge'))
-            menu.append(('q', 'Quit vimgolf'))
-            valid_codes = [x[0] for x in menu]
-            for opt in menu:
-                write('[{}] {}'.format(*opt), color='yellow')
-            selection = input_loop('Choice> ')
-            if selection not in valid_codes:
-                write('Invalid selection: {}'.format(selection), stream=sys.stderr, color='red')
-            elif selection == 'd':
-                diff_args = ['-d', '-n', infile, outfile]
-                vim(diff_args)
-            elif selection == 'w':
-                upload_status = upload_result(challenge.id, challenge.api_key, raw_keys)
-                if upload_status == Status.SUCCESS:
-                    write('Uploaded entry!', color='green')
-                    leaderboard_url = get_challenge_url(challenge.id)
-                    write('View the leaderboard: {}'.format(leaderboard_url), color='green')
-                    uploaded = True
-                    upload_eligible = False
-                else:
-                    write('The entry upload has failed', stream=sys.stderr, color='red')
-                    message = 'Please check your API key on vimgolf.com'
-                    write(message, stream=sys.stderr, color='red')
-            else:
-                break
-
-        if challenge.id:
-            challenge.add_answer(
-                keys=keycode_reprs,
-                score=score,
-                correct=correct,
-                uploaded=uploaded,
-            )
-
-        if selection == 'q':
-            break
-        write('Retrying vimgolf challenge', color='yellow')
-
-    write('Thanks for playing!', color='green')
-    return Status.SUCCESS
